Add publish and updated dates to pages.
People should be able to see when a page was published and updates, so
that they can tell if it's likely outdated or not.

It also helps automatic tools to judge the page or get extra metadata.
For example zbib.org when trying to create citing information for a page.